- Refactor API string.
The classes LipSynchronization and SpeechAnimationManager can not be tested! LipSynchronization uses a Coroutine (for a web request) which is not testable in Unit tests (and all of the other parts of LipSynchronization have been tested already). The SpeechAnimationManager can not be tested because it is a MonoBehaviour and mainly does visual rendering.